Move CreateNewIssueStep to issues UI folder
